✨ subject초기값 관련 훅으로 분리
isSubject는 나중에 스켈레톤UI관련 구현할 떄 사용할 예정입니다.
